moving codeowners.spect.ts to __tests__ (#180)
## Description
I didn't notice before, but one of the new tests introduced was placed in src/utils/ instead of __tests__

I prefer to keep all the tests grouped because when I navigate OSS projects, I often try to find the directory where tests are. This has been a great way to find out if the code has relevant tests, and learn how things are used inside the package/library.