removing mocked function from `ts-jest` (#312)
no need to see this when running tests!
<img width="1506" alt="image" src="https://user-images.githubusercontent.com/226559/173167853-ec16b65e-1750-4557-92c1-da320b2c4335.png">

I had some issues with sonarCloud so I took it out. The issues were around ignoring tests for duplication and then re-running the check. I believe some duplication in tests is necessary to understand use-cases. IMO It is documentation. Verbosity is desired in some cases.