adding Check functionality (#342)
After a bit of discussion (mainly #335) we narrowed it down into a
solution that will simplify the implementation of the action.

The functionality `--check` does something very simple: It fails if the
existing CODOWNERS file doesn't match the generated one. It will be
useful to "lint" if the committed code is in fact matching what it
should be. This functionality as said previously will simplify the
action given that many developers seem to want just the check and not to
generate the update within the workflow (Like I would have thought
initially).

I will provide documentation, but for now the code is complete and
tested (unit test)