Using swc instead of ts-jest (#348)
# Description

We always wanna go faster. That's a given. But more so when these are
tests.

I've been using SWC in other projects, and it was a good time to change.

I've also added a hook for `execute` so we can play with the `CLI`
directly.

Finally `action/setup-node` can detect the node version used from the
`.nvmrc` file. So no need to use the reusable workflow I was using
before.