Preserve block position (#347)
## Description

This PR adds a new option to preserve the position of the generated
block if the user so desires. this is helpful when there's certain
strict order the developer wants to keep. Normally because there are
manual entries.

This closes #343 

Missing:

- [x] CLI option
- [x] Documentation
- [x] Logic Refactor (because it looks ugly what I did)

Extra:

So I was playing with ChatGPT, and I decided to ask for rewrites on my
code the changes are not that bad. So I left them in