fix(action)!: Fix preserve-block-position default (#393)
Current logic adds `--preserve-block-position` to list of args if the
length of the value provided to the `preserve-block-position` github
actions variable is not zero. Since it defaults to the string `false`,
this effectively defaults it to true. You can only disable it by setting
it explicitly to an empty string. This PR corrects this logic by only
enabling `preserve-block-position` if it's explicitly set to true.

BREAKING CHANGE: Effectively change default value for
preserve-block-position in github actions