refactor: improve output messages
'Remove -check option to avoid this error' is a little ambiguous (e.g.,
will it just supress the error). Updated to let the user know what will
happen if they remove it.

Added message for when --check is successful, rather than telling the
user the file was created, which should not happen by that command.

Changed message for when new generated rules change the CO file, rather
than saying a file was created, let them know the generated rules have
changed. A CO file being created should be incredibly rare (i.e., maybe
the first time this tool is run in a project).